requirements.txt
# Core RAG Dependencies
pdfplumber
jieba
sentence-transformers
qdrant-client>=1.6.4
ollama
httpx
python-dotenv
pandas
torch
transformers
rank-bm25
beautifulsoup4
camelot-py[cv]
rapidfuzz>=3.0.0
PyMuPDF
tqdm
cn2an>=0.5.17

# Quantization Dependencies (for OOM prevention)
bitsandbytes>=0.41.0
accelerate>=0.20.0

# API Dependencies
Flask==2.3.3
Flask-CORS==4.0.0
Werkzeug==2.3.7

# Built-in modules (no need to install)
# asyncio - built-in
# threading - built-in

# LlamaIndex / GraphRAG / Neo4j
llama-index-core
llama-index-vector-stores-qdrant
llama-index-graph-stores-neo4j
neo4j
tiktoken